+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added support for declaring Angular providers via decorators
 - Added $on/$broadcast/$watch methods to $scope
 - Removed "Angular" references
+- Audited $Util classes
 
 ## [0.3.0] - 2015-07-04
 ### Fixed/Added